+++ packages/agents/lighthouse/src/bindings/backend/index.ts
@@ -1,4 +1,12 @@
-import { createLoggingContext, ExecuteArgs, jsonifyError, NxtpError, formatUrl } from "@connext/nxtp-utils";
+import {
+  createLoggingContext,
+  ExecuteArgs,
+  jsonifyError,
+  NxtpError,
+  formatUrl,
+  convertFromDbTransfer,
+  transfersCastForUrl,
+} from "@connext/nxtp-utils";
 import { constants } from "ethers";
 import axios from "axios";
 import interval from "interval-promise";
@@ -10,12 +18,17 @@ import { getContext } from "../../lighthouse";
 export const DEFAULT_SAFE_CONFIRMATIONS = 5;
 
 export const bindBackend = async (_pollInterval: number) => {
-  const { config } = getContext();
+  const { config, logger } = getContext();
+  const { requestContext, methodContext } = createLoggingContext(bindBackend.name);
   interval(async (_, stop) => {
     if (config.mode.cleanup) {
       stop();
     } else {
-      await pollBackend();
+      try {
+        await pollBackend();
+      } catch (e: unknown) {
+        logger.error("Error in pollBackend", requestContext, methodContext, jsonifyError(e as Error));
+      }
     }
   }, _pollInterval);
 };
@@ -27,51 +40,49 @@ export const pollBackend = async () => {
 
   const reconciledTransactions = await getReconciledTransactions();
 
-  reconciledTransactions.map(async (transaction: any) => {
-    // console.log(transaction);
+  await Promise.all(
+    reconciledTransactions.map(async (transaction: any) => {
+      try {
+        const xTransfer = convertFromDbTransfer(transaction);
 
-    const executeParams: ExecuteArgs = {
-      params: {
-        originDomain: transaction.origin_domain,
-        destinationDomain: transaction.destination_domain,
-        to: transaction.to,
-        callData: transaction.call_data,
-        callback: transaction.callback || constants.AddressZero,
-        callbackFee: transaction.callback_fee || "0",
-        receiveLocal: transaction.receive_local || false,
-        forceSlow: transaction.force_slow || false,
-        recovery: transaction.recovery || transaction.to,
-      },
-      local: transaction.destination_local_asset,
-      routers: [],
-      routerSignatures: [],
-      amount: transaction.destination_local_amount.toString(),
-      nonce: transaction.nonce,
-      originSender: transaction.xcall_caller,
-      relayerFee: transaction.relayer_fee || "0",
-    };
+        const executeParams: ExecuteArgs = {
+          params: {
+            originDomain: xTransfer.originDomain,
+            destinationDomain: xTransfer.destinationDomain!,
+            to: xTransfer.xparams!.to,
+            callData: xTransfer.xparams!.callData,
+            callback: xTransfer.xparams?.callback || constants.AddressZero,
+            callbackFee: xTransfer.xparams?.callbackFee || "0",
+            receiveLocal: xTransfer.xparams?.receiveLocal || false,
+            forceSlow: xTransfer.xparams?.forceSlow || false,
+            recovery: xTransfer.xparams?.recovery || xTransfer.xparams!.to,
+          },
+          local: xTransfer.destination!.assets.local.asset,
+          routers: [],
+          routerSignatures: [],
+          amount: xTransfer.destination!.assets.local.amount.toString(),
+          nonce: xTransfer.nonce!,
+          originSender: xTransfer.origin!.xcall.caller,
+          relayerFee: xTransfer.origin?.xcall.relayerFee || "0",
+        };
 
-    const transferId = transaction.trasfer_id as string;
-    try {
-      await execute(executeParams, transferId, requestContext);
-    } catch (error: any) {
-      logger.error("Error Backend Binding", requestContext, methodContext, jsonifyError(error as NxtpError), {
-        executeParams,
-        transferId,
-      });
-    }
-  });
+        const transferId = xTransfer.transferId;
+
+        await execute(executeParams, transferId, requestContext);
+      } catch (error: any) {
+        logger.error("Error Backend Binding", requestContext, methodContext, jsonifyError(error as NxtpError), {
+          transaction,
+        });
+      }
+    }),
+  );
 };
 
 export const getReconciledTransactions = async (): Promise<any> => {
   const { requestContext, methodContext } = createLoggingContext(getReconciledTransactions.name);
-  const {
-    // adapters: {},
-    logger,
-    config,
-  } = getContext();
+  const { logger, config } = getContext();
 
-  const statusIdentifier = `status=eq.Reconciled&`;
+  const statusIdentifier = `status=eq.Reconciled&${transfersCastForUrl}`;
   const uri = formatUrl(config.backendUrl, "transfers?", statusIdentifier);
   try {
     const response = await axios.get(uri);